Split services.go into services.go, internal_services.go and server.go.
